fix: adds typeguards to core and reexports from React SDK
