+++ examples/binding-to-reference-fields/resolutionUtils/resolution.ts
@@ -1,7 +1,6 @@
 import type { Entry, Asset, UnresolvedLink } from 'contentful';
-import { inMemoryEntities } from '@contentful/experiences-sdk-react';
-import { isLink, isArrayOfLinks, isAsset, isEntry } from './typeGuards';
-// TODO: maybe we can return some kind of calculation, as to - how many links failed to resolve?
+import { inMemoryEntities, isAsset, isEntry, isArrayOfLinks, isLink } from '@contentful/experiences-sdk-react';
+
 export const resolveEntityReferences = (
   entity: Entry | Asset | undefined,
 ): Entry | Asset | undefined => {
